requirements.txt

asgiref==3.11.0

certifi==2025.7.9


dj-database-url==2.3.0
Django==6.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.1

gunicorn==23.0.0

idna==3.10

olefile==0.47
packaging==25.0

pillow==11.3.0

psycopg2-binary==2.9.11

python-dotenv==1.2.1
p
requests==2.32.5

six==1.17.0

sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0

